selenium_trial/requirements.txt
selenium>=4.16.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
